requirements.txt
rich==14.2.0
jsonschema==4.25.1
numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
matplotlib==3.10.7
statsmodels==0.14.5
torch==2.9.0
pingouin==0.5.5
onnx==1.19.1
onnxscript==0.5.6
onnxruntime==1.23.2
pillow==12.0.0
torchvision==0.24.0
torch-pruning==1.6.1
questionary==2.1.1
tqdm
oapackage==2.7.16
opencv-python==4.12.0.88
dotenv==0.9.9
python-dotenv==1.2.1